requirements.txt
keras==2.15.0
matplotlib==3.8.1
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.8.1.78
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.3.2
scikit-image==0.20.0
tensorflow==2.15.0
torch==2.2.1
tqdm==4.66.4
transformers==4.40.2